Springified the logging service
