--- repose-aggregator/core/core-lib/src/main/java/com/rackspace/papi/service/logging/common/LogFrameworks.java
@@ -1,9 +1,0 @@
-package com.rackspace.papi.service.logging.common;
-
-/**
- * @author fran
- */
-public enum LogFrameworks {
-    LOG4J,
-    LOGBACK
-}
