--- repose-aggregator/core/core-lib/src/main/java/com/rackspace/papi/service/logging/facade/LoggingConfigurationFacade.java
@@ -1,10 +1,0 @@
-package com.rackspace.papi.service.logging.facade;
-
-import java.util.Properties;
-
-/**
- * @author fran
- */
-public interface LoggingConfigurationFacade {
-    void configure(Properties configurationProperties);
-}
