+++ repose-aggregator/core/core-lib/src/main/resources/spring/repose-core-context.xml
@@ -43,21 +43,10 @@
         <constructor-arg index="5" ref="threadingServiceContext"/>
     </bean>
 
-    <bean name="loggingServiceContext" class="com.rackspace.papi.service.context.impl.LoggingServiceContext">
-        <constructor-arg index="0" ref="log4jLoggingService"/>
-        <constructor-arg index="1" ref="serviceRegistry"/>
-        <constructor-arg index="2" ref="configurationManager"/>
-    </bean>
-
     <!-- Services -->
     <bean name="artifactManager" class="com.rackspace.papi.service.deploy.ArtifactManager">
         <constructor-arg index="0" ref="containerConfigurationListener"/>
     </bean>
-
-    <bean name="log4jLoggingService" class="com.rackspace.papi.service.logging.LoggingServiceImpl">
-        <constructor-arg index="0" value="LOG4J"/>
-    </bean>
-
 
     <bean name="healthCheckServiceContext"
           class="com.rackspace.papi.service.context.impl.HealthServiceContext"/>
