+++ repose-aggregator/core/valve/src/main/java/com/rackspace/cloud/valve/controller/service/context/impl/ReposeValveControllerContextManager.java
@@ -33,7 +33,6 @@ public class ReposeValveControllerContextManager implements ServletContextListen
 
       ca.getContext(EventManagerServiceContext.class).contextInitialized(sce);
       ca.getContext(ConfigurationServiceContext.class).contextInitialized(sce);
-      ca.getContext(LoggingServiceContext.class).contextInitialized(sce);
       ca.getContext(ReportingServiceContext.class).contextInitialized(sce);
       ca.getContext(ReposeValveControllerContext.class).contextInitialized(sce);
